+++ packages/style-system-demo/src/grid/col/ResponsiveGridCol.tsx
@@ -10,12 +10,12 @@ const ResponsiveGridCol: React.FC = () => (
     </p>
     <Grid spacing="0.5rem">
       <ColRow>
-        <Col size={[6, 4, 2]} dataTest="col1"><OrangeBox/></Col>
-        <Col size={[6, 4, 2]} dataTest="col2"><YellowBox/></Col>
-        <Col size={[6, 4, 2]} dataTest="col3"><OrangeBox/></Col>
-        <Col size={[6, 4, 2]} dataTest="col4"><YellowBox/></Col>
-        <Col size={[6, 4, 2]} dataTest="col5"><OrangeBox/></Col>
-        <Col size={[6, 4, 2]} dataTest="col6"><YellowBox/></Col>
+        <Col size={[6, 4, 2]} testId="col1"><OrangeBox/></Col>
+        <Col size={[6, 4, 2]} testId="col2"><YellowBox/></Col>
+        <Col size={[6, 4, 2]} testId="col3"><OrangeBox/></Col>
+        <Col size={[6, 4, 2]} testId="col4"><YellowBox/></Col>
+        <Col size={[6, 4, 2]} testId="col5"><OrangeBox/></Col>
+        <Col size={[6, 4, 2]} testId="col6"><YellowBox/></Col>
       </ColRow>
     </Grid>
     <Code>{codeString}</Code>