+++ packages/style-system/src/components/grid/Col.tsx
@@ -16,7 +16,7 @@ interface Props<T extends BaseTheme> extends WithStyle<T>,
   ResponsiveProps<StylingProps<T>>,
   ResponsiveProps<FontProps<T>> {
   size?: Size,
-  dataTest?: string,
+  testId?: string,
 }
 
 interface ColContainerProps<T extends BaseTheme> extends Props<T> {
@@ -52,7 +52,7 @@ export class Col<T extends BaseTheme> extends React.Component<Props<T>> {
   }
 
   public render() {
-    const {children, style, dataTest, ...otherProps} = this.props
+    const {children, style, testId, ...otherProps} = this.props
     const styles = mergeThemedStyles<T, Props<T>>(themedBoxStyles, style)
     return (
       <GridContext.Consumer>
@@ -60,7 +60,7 @@ export class Col<T extends BaseTheme> extends React.Component<Props<T>> {
           const child = (
             <FelaComponent<T, ColContainerProps<T>> style={styles} spacing={spacing} {...otherProps}>
               {({className}) => (
-                <div className={className} data-test={dataTest}>
+                <div className={className} data-testid={testId}>
                   {children}
                 </div>
               )}