+++ packages/style-system/src/components/grid/ColRow.tsx
@@ -15,7 +15,7 @@ import {testGridContext} from './testGridContext'
 interface Props<T extends BaseTheme> extends WithStyle<T>,
   ResponsiveProps<PaddingProps>,
   ResponsiveProps<StylingProps<T>> {
-  dataTest?: string,
+  testId?: string,
 }
 
 interface RowContainerProps<T extends BaseTheme> extends Props<T> {
@@ -155,10 +155,10 @@ class RowContainer<T extends BaseTheme> extends React.Component<RowContainerProp
         marginTop: 0,
       },
     })
-    const {style, dataTest, ...otherProps} = this.props
+    const {style, testId, ...otherProps} = this.props
     const styles = mergeThemedStyles<T, RowContainerProps<T>>(rowStyle, style)
     const renderCols = ({className, theme}: RenderProps<T>) => (
-      <div className={className} data-test={dataTest}>
+      <div className={className} data-testid={testId}>
         {rewriteCols(sortBreakpoints(theme.breakpoints), this.props.children, this.props.spacing)}
       </div>
     )